Refactor: Changes many classes and modules names, still not tested
